+++ lib/internal/bootstrap/node.js
@@ -42,8 +42,6 @@ const { getOptionValue } = NativeModule.require('internal/options');
 const config = internalBinding('config');
 const { deprecate } = NativeModule.require('internal/util');
 
-setupTraceCategoryState();
-
 setupProcessObject();
 
 setupGlobalProxy();
@@ -314,35 +312,6 @@ if (getOptionValue('--experimental-report')) {
   if (config.events.includes('signal')) {
     process.on(config.signal, handleSignal);
   }
-}
-
-function setupTraceCategoryState() {
-  const {
-    traceCategoryState,
-    setTraceCategoryStateUpdateHandler
-  } = internalBinding('trace_events');
-  const kCategoryAsyncHooks = 0;
-  let traceEventsAsyncHook;
-
-  function toggleTraceCategoryState() {
-    // Dynamically enable/disable the traceEventsAsyncHook
-    const asyncHooksEnabled = !!traceCategoryState[kCategoryAsyncHooks];
-
-    if (asyncHooksEnabled) {
-      // Lazy load internal/trace_events_async_hooks only if the async_hooks
-      // trace event category is enabled.
-      if (!traceEventsAsyncHook) {
-        traceEventsAsyncHook =
-          NativeModule.require('internal/trace_events_async_hooks');
-      }
-      traceEventsAsyncHook.enable();
-    } else if (traceEventsAsyncHook) {
-      traceEventsAsyncHook.disable();
-    }
-  }
-
-  toggleTraceCategoryState();
-  setTraceCategoryStateUpdateHandler(toggleTraceCategoryState);
 }
 
 function setupProcessObject() {
