+++ lib/internal/bootstrap/pre_execution.js
@@ -3,6 +3,8 @@
 const { getOptionValue } = require('internal/options');
 
 function prepareMainThreadExecution() {
+  setupTraceCategoryState();
+
   // If the process is spawned with env NODE_CHANNEL_FD, it's probably
   // spawned by our child_process module, then initialize IPC.
   // This attaches some internal event listeners and creates:
@@ -21,6 +23,34 @@ function prepareMainThreadExecution() {
   initializeDeprecations();
   initializeESMLoader();
   loadPreloadModules();
+}
+
+function setupTraceCategoryState() {
+  const {
+    traceCategoryState,
+    setTraceCategoryStateUpdateHandler
+  } = internalBinding('trace_events');
+  const kCategoryAsyncHooks = 0;
+  let traceEventsAsyncHook;
+
+  function toggleTraceCategoryState() {
+    // Dynamically enable/disable the traceEventsAsyncHook
+    const asyncHooksEnabled = !!traceCategoryState[kCategoryAsyncHooks];
+
+    if (asyncHooksEnabled) {
+      // Lazy load internal/trace_events_async_hooks only if the async_hooks
+      // trace event category is enabled.
+      if (!traceEventsAsyncHook) {
+        traceEventsAsyncHook = require('internal/trace_events_async_hooks');
+      }
+      traceEventsAsyncHook.enable();
+    } else if (traceEventsAsyncHook) {
+      traceEventsAsyncHook.disable();
+    }
+  }
+
+  toggleTraceCategoryState();
+  setTraceCategoryStateUpdateHandler(toggleTraceCategoryState);
 }
 
 // In general deprecations are intialized wherever the APIs are implemented,
@@ -150,5 +180,6 @@ module.exports = {
   prepareMainThreadExecution,
   initializeDeprecations,
   initializeESMLoader,
-  loadPreloadModules
+  loadPreloadModules,
+  setupTraceCategoryState
 };
