+++ lib/internal/main/check_syntax.js
@@ -22,14 +22,16 @@ if (process.argv[1] && process.argv[1] !== '-') {
   // Expand process.argv[1] into a full path.
   const path = require('path');
   process.argv[1] = path.resolve(process.argv[1]);
+
+  // TODO(joyeecheung): not every one of these are necessary
+  prepareMainThreadExecution();
+
   // Read the source.
   const filename = CJSModule._resolveFilename(process.argv[1]);
 
   const fs = require('fs');
   const source = fs.readFileSync(filename, 'utf-8');
 
-  // TODO(joyeecheung): not every one of these are necessary
-  prepareMainThreadExecution();
   markBootstrapComplete();
 
   checkScriptSyntax(source, filename);
