+++ lib/internal/main/worker_thread.js
@@ -6,7 +6,8 @@
 const {
   initializeDeprecations,
   initializeESMLoader,
-  loadPreloadModules
+  loadPreloadModules,
+  setupTraceCategoryState
 } = require('internal/bootstrap/pre_execution');
 
 const {
@@ -72,6 +73,8 @@ port.on('message', (message) => {
       manifestURL,
       hasStdin
     } = message;
+
+    setupTraceCategoryState();
     if (manifestSrc) {
       require('internal/process/policy').setup(manifestSrc, manifestURL);
     }
